code for benchmarking snips, redundant makefile calls removed

+++ benchmark.h
@@ -1,12 +1,11 @@
 #ifndef BENCHMARK_H
 #define BENCHMARK_H
 
-#include <iostream>
-#include <vector>
-#include <pair>
-#include <string>
+#include <bits/stdc++.h>
+#include "genome.h"
+#include "utils.h"
 
-std::vector<std::pair<int,char> > generateRandomInserts(int);
+std::vector<std::pair<long,char> > generateRandomInserts(long);
 void benchmarkInsert(std::string);
 
 #endif
